feat: :sparkles: 모바일 지도 레이아웃을 위한 구조 변경 #137
