+++ src/pages/map/[areaId].tsx
@@ -1,6 +1,6 @@
 import { useAppDispatch } from 'app/hook';
 import { wrapper } from 'app/store';
-import MapSidebar, { OptionLabel } from 'components/MapSidebar';
+import LocalShopInfo, { OptionLabel } from 'components/LocalShopInfo';
 import SEOUL_ENUM from 'constants/SeoulAreaEnum';
 import { initMap } from 'features/map/mapSlice';
 import { useGetShopByAreaQuery } from 'features/shops/shopApi';
@@ -10,6 +10,7 @@ import LeftArr from 'public/assets/ic_leftArr.svg';
 import { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 import { applyMediaQuery } from 'styles/mediaQuery';
+import Screen from 'styles/Screen';
 import { ShopAreaSortType } from 'types/shop';
 
 const parseAreaId = (areaId: string | string[] | undefined) => {
@@ -69,13 +70,24 @@ function MapWithAreaId(props: { areaId: number }) {
       </StyledGoBack>
       <MapContainer ref={mapRef}>
         {currentList && (
-          <MapSidebar
+          <Screen tablet desktop wide>
+            <LocalShopInfo
+              shopList={currentList}
+              currentOption={currentOption}
+              toggleOption={toggleOption}
+            />
+          </Screen>
+        )}
+      </MapContainer>
+      {currentList && (
+        <Screen mobile>
+          <LocalShopInfo
             shopList={currentList}
             currentOption={currentOption}
             toggleOption={toggleOption}
           />
-        )}
-      </MapContainer>
+        </Screen>
+      )}
     </StyledContainer>
   );
 }
@@ -90,12 +102,17 @@ export const getServerSideProps = wrapper.getServerSideProps(() => async (contex
 });
 
 const StyledContainer = styled.main`
+  height: 100%;
   display: flex;
   flex-direction: column;
   margin-top: 7.2rem;
 
   ${applyMediaQuery('desktop')} {
     margin-top: 5.2rem;
+  }
+
+  ${applyMediaQuery('mobile')} {
+    margin-top: 2.1rem;
   }
 `;
 
@@ -117,6 +134,18 @@ const StyledGoBack = styled.button`
   &:hover {
     transform: scale(1.1);
   }
+
+  ${applyMediaQuery('mobile')} {
+    gap: 0.9rem;
+    & > span {
+      font-size: 1rem;
+      line-height: 1.4rem;
+    }
+
+    & > svg {
+      transform: scale(0.5) translateY(12.5%);
+    }
+  }
 `;
 
 const LeftArrIC = styled(LeftArr)`
@@ -134,6 +163,12 @@ const MapContainer = styled.div`
     height: 55rem;
     margin: 3.5rem 0 10rem 0;
   }
+
+  ${applyMediaQuery('mobile')} {
+    height: 29rem;
+    margin: 1.1rem 0 0 0;
+  }
+
   position: relative;
 `;
